remove redundant compoent and memoize some calls
