+++ src/components/RegionMap/RegionMap.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import RegionMap from './RegionMap';
-import MetroAreaMap from './MetroAreaMap';
 import regions, { MetroArea, County, Region } from 'common/regions';
 
 export default {
@@ -18,7 +17,7 @@ export const NYMetroArea = () => {
   const metroArea = regions.findByFipsCode('35620') as Region;
   return (
     <Container>
-      <MetroAreaMap region={metroArea} width={300} height={300} />
+      <RegionMap region={metroArea} width={300} height={300} />
     </Container>
   );
 };
@@ -27,7 +26,7 @@ export const BostonMetroArea = () => {
   const metroArea = regions.findByFipsCode('14460') as MetroArea;
   return (
     <Container>
-      <MetroAreaMap region={metroArea} width={300} height={300} />
+      <RegionMap region={metroArea} width={300} height={300} />
     </Container>
   );
 };
@@ -36,7 +35,7 @@ export const KingCountyWA = () => {
   const county = regions.findByFipsCode('53033') as County;
   return (
     <Container>
-      <MetroAreaMap region={county} width={300} height={300} />
+      <RegionMap region={county} width={300} height={300} />
     </Container>
   );
 };
@@ -45,7 +44,7 @@ export const CascadeCountyMA = () => {
   const county = regions.findByFipsCode('30013') as County;
   return (
     <Container>
-      <MetroAreaMap region={county} width={300} height={300} />
+      <RegionMap region={county} width={300} height={300} />
     </Container>
   );
 };
